moved LiteralExpression
