+++ src_smartcontract/sc/CodeElement.cpp
@@ -56,12 +56,12 @@
 #include "sc_expression/CastExpression.h"
 #include "sc_expression/ConstructorArray.h"
 #include "sc_expression/ConstructorCall.h"
-#include "sc_expression/LiteralExpression.h"
 #include "sc_expression/ParenthesisExpression.h"
 #include "sc_expression/VariableIdentifier.h"
 #include "sc_expression/FunctionCallExpression.h"
 
 #include "sc_expression_literal/BooleanLiteral.h"
+#include "sc_expression_literal/LiteralExpression.h"
 #include "sc_expression_literal/NullLiteral.h"
 #include "sc_expression_literal/NumberLiteral.h"
