+++ src_smartcontract/sc/alinouslang.h
@@ -58,12 +58,12 @@
 #include "sc_expression/ConstructorArray.h"
 #include "sc_expression/ConstructorCall.h"
 #include "sc_expression/FunctionCallExpression.h"
-#include "sc_expression/LiteralExpression.h"
 #include "sc_expression/MemberReferenceExpression.h"
 #include "sc_expression/ParenthesisExpression.h"
 #include "sc_expression/VariableIdentifier.h"
 
 #include "sc_expression_literal/BooleanLiteral.h"
+#include "sc_expression_literal/LiteralExpression.h"
 #include "sc_expression_literal/NullLiteral.h"
 #include "sc_expression_literal/NumberLiteral.h"
