+++ src_smartcontract/sc_expression_literal/CMakeLists.txt
@@ -2,6 +2,7 @@
 
 set(__src
 	BooleanLiteral.cpp
+	LiteralExpression.cpp
 	NullLiteral.cpp
 	NumberLiteral.cpp
 )
